Separate input binding into user-editable file (#126)
* Separate input binding into separate file
Remove old script-based signal code

* Fix unreachable code

* Fix

* Fix InputBindingConfigPath unused warning

+++ src/graphics/graphics/opengl/GLModel.hh
@@ -31,7 +31,7 @@ namespace sp {
         void Draw(SceneShader *shader,
                   glm::mat4 modelMat,
                   const ecs::View &view,
-                  size_t boneCount,
+                  int boneCount,
                   const glm::mat4 *boneData) const;
 
         void AddPrimitive(GLModel::Primitive &prim);
